+++ src/main/scala/com/bundlepricing/repos/InMemoryRepository.scala
@@ -1,0 +1,13 @@
+package com.bundlepricing.repos
+
+import scala.concurrent.stm.TMap
+
+trait InMemoryRepository extends Repository {
+  self : RepoMetaData =>
+ 
+  protected val map = TMap.empty[Key, Entity]
+
+  def getByKey(key: Key): Option[Entity] = map.single.get(key)
+  def getAll: Map[Key, Entity] = map.single.toMap
+  def insert(entity: Entity): Unit = map.single.update(key(entity), entity)
+}
