+++ src/main/scala/com/bundlepricing/repos/repository.scala
@@ -1,7 +1,5 @@
 package com.bundlepricing.repos
 
-import scala.concurrent.stm.TMap
-
 trait RepoMetaData {
   type Entity <: AnyRef // Domain object with an unique identifier
   type Key <: AnyRef
@@ -18,12 +16,3 @@ trait Repository {
   def insert(entity: Entity): Unit
 }
 
-trait InMemoryRepository extends Repository {
-  self : RepoMetaData =>
- 
-  protected val map = TMap.empty[Key, Entity]
-
-  def getByKey(key: Key): Option[Entity] = map.single.get(key)
-  def getAll: Map[Key, Entity] = map.single.toMap
-  def insert(entity: Entity): Unit = map.single.update(key(entity), entity)
-}
