+++ src/main/scala/com/bundlepricing/repos/salatRepository.scala
@@ -7,6 +7,9 @@ import com.novus.salat.{ Context, grater }
 
 import scala.reflect.runtime.universe.typeOf
 
+/**
+ * Mongo implementation of Repository with Salat library
+ */
 trait SalatRepository extends Repository {
   self : SalatRepoMataData =>
