+++ src/test/scala/com/bundlepricing/legacy/BundlePriceSpecs.scala
@@ -1,9 +1,7 @@
-package com.bundlepricing.core
+package com.bundlepricing.legacy
 
 import com.bundlepricing.{TestData, UnitSpec}
 import com.bundlepricing.domains._
-import com.bundlepricing.legacy.BundlePrice;
-import com.bundlepricing.legacy.Inventory;
 import com.bundlepricing.repos._
 
 import org.scalatest.concurrent.ScalaFutures
@@ -15,12 +13,8 @@ import scala.concurrent.Await
  */
 class BundlePriceSpecs extends UnitSpec with TestData with ScalaFutures {
 
-
-  import com.bundlepricing.SampleData
-  
-
   import scala.concurrent.duration._
-import scala.concurrent.ExecutionContext.Implicits.global
+  import scala.concurrent.ExecutionContext.Implicits.global
     
   "BundlePrice" must "return optimized price" in {
     implicit val inventory = new Inventory with ItemRepoComponent with BundleRepoComponent {
