+++ src/test/scala/com/bundlepricing/legacy/InventorySpecs.scala
@@ -1,8 +1,7 @@
-package com.bundlepricing.core
+package com.bundlepricing.legacy
 
 import com.bundlepricing.{TestData, UnitSpec}
 import com.bundlepricing.domains._
-import com.bundlepricing.legacy.Inventory;
 import com.bundlepricing.repos._
 
 import org.scalatest.concurrent.ScalaFutures
@@ -15,7 +14,7 @@ import scala.concurrent.Await
 class InventorySpecs extends UnitSpec with TestData with ScalaFutures {
 
   import scala.concurrent.duration._
-import scala.concurrent.ExecutionContext.Implicits.global
+  import scala.concurrent.ExecutionContext.Implicits.global
     
   "Inventory" must "create Item" in new InventoryTestCxt {
     Given("instance of Inventory")
